refactor(FilterRangeDemo): Extract views
